server/requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.1.3
Flask-Cors==3.0.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.1
